pyproject.toml
[project]
name = "llmcall"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiofiles>=24.1.0",
    "openai>=1.97.0",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
    "tqdm>=4.67.1",
]
